@@ -1,31 +1,60 @@
 {
-  "name": "hasura-where-params",
-  "version": "1.1.6",
-  "description": "create where clause easy",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "devDependencies": {
-    "@types/is-empty": "^1.2.0",
-    "@types/jest": "^26.0.20",
-    "deepmerge": "^4.2.2",
-    "jest": "^26.6.3",
-    "ts-jest": "^26.4.4",
-    "typescript": "^4.2.4"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc --project tsconfig.build.json",
-    "test": "jest"
-  },
-  "author": "Paweł Mikłas",
+  "version": "0.1.0",
   "license": "MIT",
-  "dependencies": {
-    "is-empty": "^1.2.0"
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/snaqu/hasura-where-params.git"
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint",
+    "prepare": "tsdx build",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
+  },
+  "peerDependencies": {},
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  },
+  "name": "hasura-where-params",
+  "author": "vapaus",
+  "module": "dist/hasura-where-params.esm.js",
+  "size-limit": [
+    {
+      "path": "dist/hasura-where-params.cjs.production.min.js",
+      "limit": "10 KB"
+    },
+    {
+      "path": "dist/hasura-where-params.esm.js",
+      "limit": "10 KB"
+    }
+  ],
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^4.12.0",
+    "husky": "^6.0.0",
+    "size-limit": "^4.12.0",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.3.0",
+    "typescript": "^4.3.4"
+  },
+  "dependencies": {
+    "@types/is-empty": "^1.2.0",
+    "deepmerge": "^4.2.2",
+    "is-empty": "^1.2.0"
   }
 }
