@@ -1,9 +1,8 @@
-import type * as DeepMergeType from "deepmerge";
-const deepMerge: typeof DeepMergeType = require("deepmerge");
-import { truthy } from "./utils/truthy";
-import empty from "is-empty";
+import deepMerge from 'deepmerge';
+import { truthy } from './utils/truthy';
+import empty from 'is-empty';
 
-const ARRAY_FIELDS = ["_and", "_or"];
+const ARRAY_FIELDS = ['_and', '_or'];
 
 const whereArgs = <T extends object>(...filters: (Object | undefined)[]): T =>
   deepMerge.all(filters.filter(truthy)) as T;
@@ -18,7 +17,7 @@ const whereFilter = (
   }
 
   if ((condition == null && !empty(value)) || condition === true) {
-    return path.split(".").reduceRight((acc, item) => {
+    return path.split('.').reduceRight((acc, item) => {
       if (ARRAY_FIELDS.includes(item)) {
         acc = { [item]: [{ ...acc }] };
       } else {
